// spaceship-titanic/src/app/models/passenger.model.ts
export interface Passenger {
  // PassengerId: string;
  HomePlanet: 'Earth' | 'Europa' | 'Mars';
  CryoSleep: boolean;
  // Cabin: string;
  Destination: 'TRAPPIST-1e' | '55 Cancri e' | 'PSO J318.5-22';
  Age: number;
  VIP: boolean;
  // Name: string;
  // Transported: boolean;
  RoomService: number;
  FoodCourt: number;
  ShoppingMall: number;
  Spa: number;
  VRDeck: number;
}